% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fuzzyforest.R
\docType{data}
\name{ctg}
\alias{ctg}
\title{Cardiotocography Data Set}
\format{A data frame with 100 rows and 21.}
\usage{
data(ctg)
}
\description{
A data set containing measurements of fetal heart rate and uterine
contraction from cardiotocograms.  This data set was obtained from
the [UCI machine learning repository](https://archive.ics.uci.edu/ml/index.html)
For our examples we extract a random sub sample of 100 observations.
}
\keyword{datasets}

