% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fuzzyforest.R
\docType{data}
\name{example_ff}
\alias{example_ff}
\title{Fuzzy Forest Example}
\format{.RData}
\description{
An example of a fuzzy_forest object derived from fitting fuzzy forests
on the ctg data set.  The source code used to produce example_ff can be
seen in the vignette "fuzzyforest_introduction".
}
\keyword{R}
\keyword{object}

