% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fuzzy_forest_obj.R
\name{fuzzy_forest}
\alias{fuzzy_forest}
\title{Fuzzy Forest Object}
\usage{
fuzzy_forest(feature_list, final_rf, module_membership, WGCNA_object = NULL,
  survivor_list, selection_list)
}
\arguments{
\item{feature_list}{List of selected features along with variable
importance measures.}

\item{final_rf}{A final random forest fit using the features
selected by fuzzy forests.}

\item{module_membership}{Module membership of each feature.}

\item{WGCNA_object}{If applicable, output of WGCNA analysis.}

\item{survivor_list}{List of features that have survived screening step.}

\item{selection_list}{List of features retained at each iteration of
selection step.}
}
\value{
An object of type fuzzy_forest.
}
\description{
Fuzzy forests returns an object of type
fuzzyforest.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

