% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fuzzyforest}
\alias{fuzzyforest}
\title{Fits fuzzy forest algorithm.}
\usage{
fuzzyforest(X, y, module_membership, proportion_kept, screening_mtry,
  selection_mtry)
}
\arguments{
\item{X}{A data.frame with feature vectors.}

\item{y}{Response vector.}

\item{module_membership}{A data.frame giving the module membership of each
feature.
The first column gives the name of feature.
The second column gives the name of partition.}

\item{proportion_kept}{A number between 0 and 1. The percent of features
kept at each iteration of random forest during the
screening step.}

\item{screening_mtry}{A number between 0 and 1.  Mtry for each random forest
is set to \code{p_current*screening_mtry} where
\code{p_current} is the current number of features.}

\item{selection_mtry}{A number betwee 0 and 1.  Mtry for final selection
step is set to \code{num_survivors*selection_mtry}
where \code{num_survivors} is the number features
that have survived the initial screening step.}
}
\description{
Fits fuzzy forest algorithm.  Returns
fuzzy forest object.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

