% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{fuzzyforests}
\alias{fuzzyforests}
\alias{fuzzyforests-package}
\title{fuzzyforest: an implementation of the fuzzy forest algorithm in R.}
\description{
This package implements fuzzy forests and integrates the fuzzy
forests algorithm with the package, \pkg{WGCNA}.
}

