% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fuzzyforest.R
\docType{data}
\name{Liver_Expr}
\alias{Liver_Expr}
\title{Liver Expression Data from Female Mice}
\format{A data frame with 66 rows and 3601}
\usage{
data(Liver_Expr)
}
\description{
A data set containing gene expression levels in liver tissue from female
mice. This data set is a subset of the liver expression data set
from the WGCNA tutorial \url{http://labs.genetics.ucla.edu/horvath/CoexpressionNetwork/Rpackages/WGCNA/Tutorials/}.
The tutorial contains further information about the data set as well as
extensive examples of WGCNA.
}
\details{
\itemize{
  \item The first column contains weight (g) for the 66 mice.
  \item The other 3600 columns contain the liver expression levels.
 }
}
\keyword{datasets}

