% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fuzzy_forest_obj.R
\name{modplot}
\alias{modplot}
\title{Plots relative importance of modules.}
\usage{
modplot(object, main = NULL, xlab = NULL, ylab = NULL,
  module_labels = NULL)
}
\arguments{
\item{object}{A fuzzy_forest object.}

\item{main}{Title of plot.}

\item{xlab}{Title for the x axis.}

\item{ylab}{Title for the y axis.}

\item{module_labels}{Labels for the modules.  A data.frame
or character matrix with first column giving
the current name of module and second column giving
the assigned name of each module.}
}
\description{
The plot is designed
to depict the size of each module and what percentage of selected
features fall into each module.  In particular, it is easy to
determine which module is over-represented in the group of selected
features.
}

