% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fuzzy_forest_obj.R
\name{plot.fuzzy_forest}
\alias{plot.fuzzy_forest}
\title{Plot method for fuzzy_forest object.
Plots results of fuzzy forest algorithm.}
\usage{
\method{plot}{fuzzy_forest}(x, ...)
}
\arguments{
\item{x}{A fuzzy_forest object.}

\item{...}{Additional arguments not in use.}
}
\description{
Plot method for fuzzy_forest object.
Plots results of fuzzy forest algorithm.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

