% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fuzzy_forest_obj.R
\name{predict.fuzzy_forest}
\alias{predict.fuzzy_forest}
\title{Predict method for fuzzy_forest object.
Obtains predictions from fuzzy forest algorithm.}
\usage{
\method{predict}{fuzzy_forest}(fuzzy_forest, new_data)
}
\arguments{
\item{fuzzy_forest}{A fuzzy_forest object.}

\item{new_data}{A matrix or data.frame containing new_data.
                Pay close attention to ensure feature names
                match between training set and test set
                data.frame.}
}
\value{
A vector of predictions
}
\description{
Predict method for fuzzy_forest object.
Obtains predictions from fuzzy forest algorithm.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

