% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fuzzy_forest_obj.R
\name{print.fuzzy_forest}
\alias{print.fuzzy_forest}
\title{Print fuzzy_forest object.
Prints output from fuzzy forests algorithm.}
\usage{
\method{print}{fuzzy_forest}(x, ...)
}
\arguments{
\item{x}{A fuzzy_forest object.}

\item{...}{additional arguments}
}
\value{
data.frame with list of selected features and variable
         importance measures.
}
\description{
Print fuzzy_forest object.
Prints output from fuzzy forests algorithm.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

