% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fuzzy_forest_obj.R
\name{relabel_modules}
\alias{relabel_modules}
\title{Relabel modules.}
\usage{
relabel_modules(object, module_labels)
}
\arguments{
\item{object}{An object of type fuzzy_forest.}

\item{module_labels}{Labels for the modules.  A data.frame
or character matrix with first column giving
the current name of module and second column giving
the assigned name of each module.}
}
\description{
Lets user easily re-label modules.  Modules are often labeled according to
color in WGCNA.  This function allows the user to rename the modules.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

