% Generated by roxygen2 (4.0.2): do not edit by hand
\name{screen_control}
\alias{screen_control}
\title{Set Parameters for Screening Step of
Fuzzy Forests}
\usage{
screen_control(drop_fraction = 0.25, stop_fraction = 0.05,
  mtry_factor = 1, min_ntree = 10000, ntree_factor = 10)
}
\arguments{
\item{drop_fraction}{A number between 0 and 1.  Percentage of features
dropped at each iteration.}

\item{stop_fraction}{A number between 0 and 1. Proportion features
from each module to retain at screening step.}

\item{mtry_factor}{\code{mtry} is set to
\eqn{\sqrt(p)}*\code{mtry_factor}.}

\item{min_ntree}{Set minimum value for \code{ntree} in each random
forest}

\item{ntree_factor}{A number greater than 1.  \code{ntree} for each
random is \code{ntree_factor} times the number
of features.}
}
\value{
An object of type screen_control.
}
\description{
Creates \code{screen_control} object for
controlling how feature selection
will be carried out on each module.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

