% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fuzzyforest_fit.R
\name{wff}
\alias{wff}
\title{Fits WGCNA based fuzzy forest algorithm.}
\usage{
wff(X, y, Z = NULL, WGCNA_params = WGCNA_control(p = 6),
  screen_params = screen_control(min_ntree = 5000),
  select_params = select_control(min_ntree = 5000), final_ntree = 500,
  num_processors = 1, nodesize, test_features = NULL, test_y = NULL)
}
\arguments{
\item{X}{A data.frame. Each column corresponds to a feature
vector.  WGCNA will be used to cluster the
features in X.  As a result, the features should be
all be numeric.  Non-numeric features may be input
via Z.}

\item{y}{Response vector.  For classification, y should be a
factor or a character.  For regression, y should be
numeric.}

\item{Z}{Additional features that are not to be screened out
at the screening step.  WGCNA is not carried out on
features in Z.}

\item{WGCNA_params}{Parameters for WGCNA.
See \code{\link[WGCNA]{blockwiseModules}} and
\code{\link[fuzzyforest]{WGCNA_control}} for details.
\code{WGCNA_params} is an object of type
\code{WGCNA_control}.}

\item{screen_params}{Parameters for screening step of fuzzy forests.
See \code{\link[fuzzyforest]{screen_control}} for details.
\code{screen_params} is an object of type
\code{screen_control}.}

\item{select_params}{Parameters for selection step of fuzzy forests.
See \code{\link[fuzzyforest]{select_control}} for details.
\code{select_params} is an object of type
\code{select_control}.}

\item{final_ntree}{Number trees grown in the final random forest.
This random forest contains all selected features.}

\item{num_processors}{Number of processors used to fit random forests.}

\item{nodesize}{Minimum terminal nodesize. 1 if classification.
5 if regression.  If the sample size is very large,
the trees will be grown extremely deep.
This may lead to issues with memory usage and may
lead to significant increases in the time it takes
the algorithm to run.  In this case,
it may be useful to increase \code{nodesize}.}

\item{test_features}{A data.frame containing features from a test set.
The data.frame should contain the features in both
X and Z.}

\item{test_y}{The responses for the test set.}
}
\value{
An object of type \code{\link[fuzzyforest]{fuzzy_forest}}.  This
object is a list containing useful output of fuzzy forests.
In particular it contains a data.frame with list of selected features.
It also includes the random forest fit using the selected features.
}
\description{
Fits fuzzy forest algorithm using WGCNA.  Returns
fuzzy forest object.
}
\note{
This work was partially funded by NSF IIS 1251151.
}
\examples{
library(WGCNA)
library(randomForest)
data(Liver_Expr)
set.seed(3653)
y <- Liver_Expr$weight
X <- Liver_Expr[, -1]
screen_params <- screen_control(min_ntree=5000, keep_fraction=.25)
select_params <- select_control(min_ntree=5000, number_selected=10)
# fit fuzzy forests
wff_fit <- wff(X, y, WGCNA_params = WGCNA_control(p=8),
               screen_params = screen_params,
               select_params = select_params,
               final_ntree = 5000)

#see which features have been selected along with their variable importance
feature_list <- wff_fit$feature_list

#see which modules are over represented in terms of their importance
modplot(wff_fit)

#produce dotplot to obtain a visual representation of the variable importance measures.
varImpPlot(wff_fit$final_rf)
}

