% Generated by roxygen2 (4.0.2): do not edit by hand
\name{WGCNA_control}
\alias{WGCNA_control}
\title{Set Parameters for WGCNA Step of
Fuzzy Forests}
\usage{
WGCNA_control(power = 6, ...)
}
\arguments{
\item{power}{Power of adjacency function.}

\item{...}{Other arguments.  See documentation for
WGCNA.}
}
\value{
An object of type WGCNA_control.
}
\description{
Creates \code{WGCNA_control} object for
controlling WGCNA will be carried out.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

