% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fuzzy_forest_obj.R
\name{fuzzy_forest}
\alias{fuzzy_forest}
\title{Fuzzy Forest Object}
\usage{
fuzzy_forest(feature_list, final_rf, module_membership, WGCNA_object = NULL)
}
\arguments{
\item{feature_list}{List of selected features with variable
importance.}

\item{final_rf}{The final random forest that was fit by
fuzzy forests.}

\item{module_membership}{Module membership of each feature.}

\item{WGCNA_object}{output of WGCNA analysis.}
}
\value{
An object of type fuzzy_forest.
}
\description{
Fuzzy forests returns an object of type
fuzzyforest.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

