% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fuzzyforest}
\alias{fuzzyforest}
\title{Fits fuzzy forest algorithm.}
\usage{
fuzzyforest(X, y, module_membership, proportion_dropped, proportion_kept,
  screening_mtry, selection_mtry)
}
\arguments{
\item{X}{A data.frame.  Each column corresponds to a feature vectors.}

\item{y}{Response vector.}

\item{module_membership}{A data.frame with number of rows equal to
\code{ncol(X)}.
The first column gives the name of module.
The second column gives the module membership
of feature.}

\item{proportion_dropped}{A number between 0 and 1.  The proportion dropped
at each stage of the iterative random forests.}

\item{proportion_kept}{A number between 0 and 1. The percent of features
kept at each iteration of random forest during the
screening step.}

\item{screening_mtry}{A number between 0 and 1.  Mtry for each random forest
is set to \code{p_current*screening_mtry} where
\code{p_current} is the current number of features.}

\item{selection_mtry}{A number betwee 0 and 1.  Mtry for final selection
step is set to \code{num_survivors*selection_mtry}
where \code{num_survivors} is the number features
that have survived the initial screening step.}

\item{ntree_factor}{A number greater than 1.  \code{ntree} for each random
is \code{ntree_factor} times the number of features.}
}
\description{
Fits fuzzy forest algorithm.  Returns
fuzzy forest object.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

