% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fuzzyforest.R
\docType{package}
\name{fuzzyforest}
\alias{fuzzyforest}
\alias{fuzzyforest-package}
\title{fuzzyforest: an implementation of the fuzzy forest algorithm in R.}
\description{
This package implements fuzzy forests and integrates the fuzzy
forests algorithm with the package, \pkg{WGCNA}.
}

