% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fuzzyforest}
\alias{fuzzyforest}
\title{Fits fuzzy forest algorithm.}
\usage{
fuzzyforest(X, y, module_membership, screen_control = screen_control(),
  select_control = select_control(), num_processors = 1)
}
\arguments{
\item{X}{A data.frame.
Each column corresponds to a feature vectors.}

\item{y}{Response vector.}

\item{module_membership}{A vector giving module membership of each feature.}

\item{screen_control}{A screen control object.}

\item{select_control}{A selection control object}

\item{num_processors}{Number of processors used to fit random forests.}
}
\value{
A data.frame with the top ranked features.
}
\description{
Fits fuzzy forest algorithm.  Returns
fuzzy forest object.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

