% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fuzzyforest}
\alias{fuzzyforest}
\title{Fits fuzzy forest algorithm.}
\usage{
fuzzyforest(X, y, module_membership, drop_fraction = 0.25,
  stop_fraction = 0.05, mtry_factor = 1, ntree_factor = 10,
  min_ntree = 5000, num_processors = 1)
}
\arguments{
\item{X}{A data.frame.
Each column corresponds to a feature vectors.}

\item{y}{Response vector.}

\item{module_membership}{A data.frame with number of rows equal to
\code{ncol(X)}.
The first column gives the name of module.
The second column gives the module membership
of feature.}

\item{drop_fraction}{A number between 0 and 1.  Percentage of features
dropped at each iteration.}

\item{stop_fraction}{A number between 0 and 1. Proportion features
from each module to retain at screening step.}

\item{mtry_factor}{A positive number.  Mtry for each random forest
is set to
\code{ceiling}(\eqn{\sqrt{p}}\code{mtry_factor})
where \code{p} is the current number of features.}

\item{ntree_factor}{A number greater than 1.  \code{ntree} for each
random is \code{ntree_factor} times the number
of features.  For each random forest, \code{ntree}
is set to \code{max}(\code{min_ntree},
\code{ntree_factor}*\code{p}).}

\item{min_ntree}{Minimum number of trees grown in each random forest.}

\item{num_processors}{Number of processors used to fit random forests.}
}
\value{
A data.frame with the top ranked features.
}
\description{
Fits fuzzy forest algorithm.  Returns
fuzzy forest object.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

