% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iterative_RF}
\alias{iterative_RF}
\title{Fits iterative random forest algorithm.}
\usage{
iterative_RF(X, y, drop_fraction, stop_fraction, mtry_fraction,
  ntree_factor = 10, num_processors = 2)
}
\arguments{
\item{X}{A data.frame.
Each column corresponds to a feature vectors.}

\item{y}{Response vector.}

\item{drop_fraction}{A number between 0 and 1.  Percentage of features
dropped at each iteration.}

\item{stop_fraction}{A number between 0 and 1. Proportion features
from each module to retain at screening step.}

\item{mtry_fraction}{A number between 0 and 1.  Mtry for each random forest
is set to \code{p_current*mtry_fraction} where
\code{p_current} is the current number of features.}

\item{ntree_factor}{A number greater than 1.  \code{ntree} for each random
is \code{ntree_factor} times the number of features.}

\item{num_processors}{Number of processors used to fit random forests.}

\item{module_membership}{A data.frame with number of rows equal to
\code{ncol(X)}.
The first column gives the name of module.
The second column gives the module membership
of feature.}
}
\description{
Fits iterative random forest algorithm.  Returns
data.frame with variable importances and top rated features.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

