% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tuning_parameters.R
\name{screen_control}
\alias{screen_control}
\title{Set Parameters for Screening Step of
Fuzzy Forests}
\usage{
screen_control(drop_fraction = 0.25, keep_fraction = 0.05,
  mtry_factor = 1, min_ntree = 10000, ntree_factor = 10)
}
\arguments{
\item{drop_fraction}{A number between 0 and 1.  Percentage of features
dropped at each iteration.}

\item{keep_fraction}{A number between 0 and 1. Proportion features
from each module to retain at screening step.}

\item{mtry_factor}{\code{mtry} is set to
\eqn{\sqrt(p)}*\code{mtry_factor}.}

\item{min_ntree}{Minimum number of trees grown in each random forest.}

\item{ntree_factor}{A number greater than 1.  \code{ntree} for each
random is \code{ntree_factor} times the number
of features.  For each random forest, \code{ntree}
is set to \code{max}(\code{min_ntree},
\code{ntree_factor}*\code{p}).}
}
\value{
An object of type screen_control.
}
\description{
Creates \code{screen_control} object for
controlling how feature selection
will be carried out on each module.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

