% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_control}
\alias{select_control}
\title{Set Parameters for Selection Step of
Fuzzy Forests}
\usage{
select_control(drop_fraction = 0.25, number_selected = 5, mtry_factor = 1,
  min_ntree = 10000, ntree_factor = 10)
}
\arguments{
\item{drop_fraction}{A number between 0 and 1.  Percentage of features
dropped at each iteration.}

\item{number_selected}{A positive number. Number of features
that will be selected by fuzzyforests.}

\item{mtry_factor}{A number between 0 and 1.  Mtry for each random forest
is set to \code{p_current*mtry_fraction} where
\code{p_current} is the current number of features.}

\item{min_ntree}{Set minimum value for \code{ntree} in each random
forest}

\item{ntree_factor}{A number greater than 1.  \code{ntree} for each
random is \code{ntree_factor} times the number
of features.}
}
\value{
An object of type selection_control.
}
\description{
Creates \code{selection_control} object for
controlling how feature selection
will be carried out after features from different
modules have been combined.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

