% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fuzzyforest_fit.R
\name{WGCNA_fuzzyforest}
\alias{WGCNA_fuzzyforest}
\title{Fits WGCNA based fuzzy forest algorithm.}
\usage{
WGCNA_fuzzyforest(X, y, WGCNA_params = WGCNA_control(p = 6),
  screen_params = screen_control(min_ntree = 5000),
  select_params = select_control(min_ntree = 5000), num_processors = 1)
}
\arguments{
\item{X}{A data.frame.
Each column corresponds to a feature vectors.}

\item{y}{Response vector.}

\item{WGCNA_params}{Parameters for WGCNA.
See \code{\link[WGCNA]{blockwiseModules}} and
\code{\link[fuzzyforest]{WGCNA_control}} for details.
\code{WGCNA_params} is an object of type
\code{WGCNA_control}.}

\item{screen_params}{Parameters for screening step of fuzzy forests.
See \code{\link[fuzzyforest]{screen_control}} for details.
\code{screen_params} is an object of type
\code{screen_control}.}

\item{select_params}{Parameters for selection step of fuzzy forests.
See \code{\link[fuzzyforest]{select_control}} for details.
\code{select_params} is an object of type
\code{select_control}.}

\item{num_processors}{Number of processors.}
}
\value{
A data.frame with the top ranked features.
}
\description{
Fits fuzzy forest algorithm using WGCNA.  Returns
fuzzy forest object.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

