% Generated by roxygen2 (4.0.2): do not edit by hand
\name{WGCNA_fuzzyforest}
\alias{WGCNA_fuzzyforest}
\title{Fits WGCNA based fuzzy forest algorithm.}
\usage{
WGCNA_fuzzyforest(X, y, number_selected, WGCNA_control = WGCNA_control(power =
  6), screen_control = NULL, select_control = NULL, num_processors = NULL)
}
\arguments{
\item{X}{A data.frame.
Each column corresponds to a feature vectors.}

\item{y}{Response vector.}

\item{number_selected}{Number of features selected by algorithm.}

\item{WGCNA_control}{An object of class WGCNA_control.}

\item{screen_control}{An object of class screen_control.}

\item{select_control}{An object of class selection_control.}

\item{num_processors}{Number of processors.}
}
\value{
A data.frame with the top ranked features.
}
\description{
Fits fuzzy forest algorithm using WGCNA.  Returns
fuzzy forest object.
}
\note{
This work was partially funded by NSF IIS 1251151.
}

